requirements.txt
amqp==5.3.1
asgiref==3.11.1
asttokens==3.0.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
decorator==5.2.1
Django==5.2
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-environ==0.13.0
django-extensions==4.1
django-filter==25.2
django-redis==6.0.0
django-restframework==0.0.1
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
executing==2.2.1
factory_boy==3.3.3
Faker==40.11.0
gdal-installer==1.0.2
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
packaging==26.0
parso==0.8.6
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sentry-sdk==2.54.0
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
tqdm==4.67.3
traitlets==5.14.3
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.6
